[Feature] creating the modules from schatch but using fresh types in placed of the generated by prisma nest library
This commit adds support for filtering payment schedules by custom criteria using the `where` argument. The `where` argument is of type `PaymentScheduleWhereInput` and allows users to specify filter criteria for the `PaymentSchedule` objects to retrieve.

To implement this feature, I modified the `getPaymentSchedules` method in the `PaymentScheduleResolver` class to take an optional `where` argument of type `PaymentScheduleWhereInput`. I then used the spread operator to merge the `where` object with the existing `where` clause of the SQL query.