[FEATURE]: Add payment schedule creation to bulk loan creation
This pull request adds the ability to create payment schedules when creating loans in bulk. Previously, payment schedules had to be created manually after the loans were created, which was time-consuming and error-prone.

To solve this problem, I added a new method to the `LoanService` that creates payment schedules for each loan in the bulk creation request. I also updated the `findLoanWithAdditionalData` method to include payment schedules in the returned data.

This change required updating the `LoanService` and the `LoanResolver` to handle the new method. I also added some additional tests to ensure that the new functionality works as expected.