feat: Update loan types resolver and types
Refactor the `LoanTypesResolver` class in `loan-types.resolver.ts` to include a `where` argument in the `findMany` method. Also, update the `LoanType` class in `types.ts` to make the `availableFor` field an array of `EmployeesTypes` instead of a single value. This change allows for more flexibility in specifying the types of employees that a loan type is available for.